+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # @digitalbazaar/ed25519-verification-key-2020 ChangeLog
 
+## 2.1.0 -
+
+### Added
+- Add `revoked` export tests, `SUITE_CONTEXT` class property. (To support
+  `CryptoLD`'s new `fromKeyId()` method.)
+
 ## 2.0.0 - 2021-03-17
 
 ## Changed